MOAR TESTS
tests:
add autotest.sh to automatically compile, test and compare output with std
add rbegin and rend tests
add clear tests
add insert tests
add preserving tests for empty map
refacto main

Vector:
fix resize not destroying every discarded instances

bst:
fix clear() shitting itself (crashing) when root == NULL

+++ bits/bst.hpp
@@ -590,8 +590,11 @@ namespace ft {
 			void clear() {
 				anchor_guard_type guard(*this);
 
-				destroy_from_dyn_storage(root);
+				if (root != NULL) {
+					destroy_from_dyn_storage(root);
+				}
 				root = NULL;
+				element_cout = 0;
 			}
 
 			node_type *retrieve_node_key(const _Key &key, node_type **parent = NULL, bool *is_less = NULL) const {
